updating Detector and multipy to reduce resetting

+++ multipy_main.py
@@ -263,7 +263,10 @@ while(True):
     try:
     	with open('error.txt','r') as f:
 		for line in f:
-			if len(line.strip()) > 0:
+			#print(line)
+			contain = line.strip().split(' ')
+			if (len(line.strip()) > 0) and '(38)' not in contain and '(107)' not in contain and '(111)' not in contain and 'unlikely' not in contain and 'Traceback' not in contain:
+				print(line)
 				print('something wrong, resetting')
 				sleep(5)
 				os.remove('error.txt')
